requirements.txt
Flask>=2.1.2
lxml
pillow
requests